// radio-read-consumer/src/exceptions/dict/exception-codes.enum.ts
export enum SqlExceptionCode {
  CONNECTION_ERROR = 1000,
  DB_READ_ERROR = 1010,
  DB_INSERT_ERROR = 1020,
  DB_UPDATE_ERROR = 1030,
  DB_DELETE_ERROR = 1040,
  UNKNOWN_ERROR = 1050,
}

export enum RabbitExceptionCode {
  CONNECTION_ERROR = 2000,
  MESSAGES_READING_ERROR = 2010,
  MESSAGES_SENDING_ERROR = 2020,
  UNKNOWN_ERROR = 2040,
}

export enum ValidationExceptionCode {
  VALIDATION_ERROR = 3000,
  UNKNOWN_VALIDATION_ERROR = 3040,
}